+++ src/Service/RdsDataService/src/Exception/DatabaseErrorException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\RdsDataService\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * There was an error in processing the SQL statement.
+ */
+final class DatabaseErrorException extends ClientException
+{
+}