+++ src/Service/RdsDataService/src/Exception/ServiceUnavailableErrorException.php
@@ -5,7 +5,7 @@ namespace AsyncAws\RdsDataService\Exception;
 use AsyncAws\Core\Exception\Http\ServerException;
 
 /**
- * The service specified by the `resourceArn` parameter is not available.
+ * The service specified by the `resourceArn` parameter isn't available.
  */
 final class ServiceUnavailableErrorException extends ServerException
 {